Move decode_base64 to text module too
